Create chapter_4_hw.R
The Numbeo website (www.numbeo.com) provides access to a variety of data. One table lists prices of
certain items in selected cities around the world. They also report an overall cost-of-living index for each city
compared to the costs of hundreds of items in New York City. For example, London at 110.69 is 10.69%
more expensive than New York. In the data file Cost_of_living_2013.txt included are the Cost of Living
Index, a Rent Index, a Groceries Index, a Restaurant Price Index, and a Local Purchasing Power Index that
measures the ability of the average wage earner in a city to buy foods and services. All indices are measured
relative to New York City, which is scored 100.